Group extension point declaration + rename package per language. See https://github.com/mickaelistria/eclipse-bluesky/issues/26
